refactor: Phase 2 - Move repositories to data/repository/
- Move AIRepository, AIService, PremiumFeatureManager from ai/ to data/repository/
- Move ChatRepository from chat/ to data/repository/
- Move UserRepository from repo/ to data/repository/
- Move RevenueCatManager from subscription/ to data/repository/
- Update all package declarations and imports
