+++ app/src/main/java/com/kharagedition/tibetankeyboard/AIKeyboardView.kt
@@ -12,12 +12,12 @@ import android.view.View
 import android.widget.*
 import androidx.lifecycle.LifecycleOwner
 import com.kharagedition.tibetankeyboard.ai.AIKeyboardInterface
-import com.kharagedition.tibetankeyboard.ai.AIService
+import com.kharagedition.tibetankeyboard.data.repository.AIService
 import com.kharagedition.tibetankeyboard.auth.AuthManager
 import com.kharagedition.tibetankeyboard.data.model.GrammarResult
 import com.kharagedition.tibetankeyboard.data.model.RephraseResult
 import com.kharagedition.tibetankeyboard.data.model.TranslationResult
-import com.kharagedition.tibetankeyboard.subscription.RevenueCatManager
+import com.kharagedition.tibetankeyboard.data.repository.RevenueCatManager
 import com.kharagedition.tibetankeyboard.subscription.SubscriptionUIComponent
 import kotlinx.coroutines.*
