+++ app/src/main/java/com/kharagedition/tibetankeyboard/LoginActivity.kt
@@ -33,9 +33,9 @@ import com.google.firebase.auth.ktx.auth
 import com.google.firebase.ktx.Firebase
 import com.kharagedition.tibetankeyboard.R
 import com.kharagedition.tibetankeyboard.data.local.UserPreferences
-import com.kharagedition.tibetankeyboard.repo.UserRepository
+import com.kharagedition.tibetankeyboard.data.repository.UserRepository
 import com.kharagedition.tibetankeyboard.ui.ChatActivity
-import com.kharagedition.tibetankeyboard.subscription.RevenueCatManager
+import com.kharagedition.tibetankeyboard.data.repository.RevenueCatManager
 import com.kharagedition.tibetankeyboard.auth.AuthManager
 import android.util.Log
 import kotlinx.coroutines.launch
