+++ app/src/main/java/com/kharagedition/tibetankeyboard/auth/AuthManager.kt
@@ -10,7 +10,7 @@ import com.google.firebase.auth.ktx.auth
 import com.google.firebase.ktx.Firebase
 import com.kharagedition.tibetankeyboard.LoginActivity
 import com.kharagedition.tibetankeyboard.data.local.UserPreferences
-import com.kharagedition.tibetankeyboard.subscription.RevenueCatManager
+import com.kharagedition.tibetankeyboard.data.repository.RevenueCatManager
 
 /**
  * Manages authentication state and user session
