+++ app/src/main/java/com/kharagedition/tibetankeyboard/data/repository/UserRepository.kt
@@ -1,4 +1,4 @@
-package com.kharagedition.tibetankeyboard.repo
+package com.kharagedition.tibetankeyboard.data.repository
 
 // UserRepository.kt - Repository for user data operations
 import com.google.firebase.firestore.FirebaseFirestore
